requirements.txt
setuptools
patch-ng
anyio
addict
astunparse
blendmodes
clean-fid
filetype
future
GitPython
httpcore
inflection
jsonmerge
kornia
lark
lpips
omegaconf
open-clip-torch
optimum
piexif
psutil
pyyaml
resize-right
rich
safetensors
scipy
tensordict==0.1.2
toml
torchdiffeq
voluptuous
yapf
scikit-image
fasteners
dctorch
pymatting
peft
orjson
httpx==0.24.1
compel==2.0.2
torchsde==0.2.6
clip-interrogator==0.6.0
antlr4-python3-runtime==4.9.3
requests==2.31.0
tqdm==4.66.1
accelerate==0.28.0
opencv-contrib-python-headless==4.9.0.80
diffusers==0.27.0
einops==0.4.1
gradio==3.43.2
huggingface_hub==0.21.4
numexpr==2.8.8
numpy==1.26.4
numba==0.59.0
pandas
protobuf==3.20.3
pytorch_lightning==1.9.4
tokenizers==0.15.2
transformers==4.38.2
tomesd==0.1.3
urllib3==1.26.18
Pillow==10.2.0
timm==0.9.12
pydantic==1.10.13
typing-extensions==4.9.0
